feat: implementa result pattern

+++ src/Application/UseCases/Alunos/Delete/DeleteAlunoCommandHandler.cs
@@ -1,19 +1,22 @@
+using Application.Common;
 using Application.Interfaces;
 using MediatR;
 
 namespace Application.UseCases.Alunos;
 
-public class DeleteAlunoCommandHandler(IAlunosRepository repo) : IRequestHandler<DeleteAlunoCommand, Unit>
+public class DeleteAlunoCommandHandler(IAlunosRepository repo) : IRequestHandler<DeleteAlunoCommand, Result>
 {
     private readonly IAlunosRepository _repo = repo;
 
-    public async Task<Unit> Handle(DeleteAlunoCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeleteAlunoCommand request, CancellationToken cancellationToken)
     {
-        var aluno = await _repo.GetById(request.Id)
-            ?? throw new Exception("Aluno não encontrado");
-        
+        var aluno = await _repo.GetById(request.Id);
+
+        if (aluno == null)
+            return Error.NotFound("Aluno não encontrado");
+
         await _repo.Delete(aluno);
 
-        return Unit.Value;
+        return Result.Sucess();
     }
 }
